Rename composer/datasets/webdataset.py => composer/datasets/webdataset_utils.py
