+++ composer/datasets/webdataset_utils.py
