+++ scripts/webdatasets/ade20k.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, Iterable, List, Tuple
 
 from PIL import Image
 
-from composer.datasets.webdataset import create_webdataset
+from composer.datasets.webdataset_utils import create_webdataset
 
 
 """
